test(pdf): extract table-scope span-caption family (#250)
* test(research): extract span caption fixtures

* test(research): extract weak slab fixtures

* test(research): remove unused scope import

* test(research): extract caption lane fixtures

* test(research): extract source completion fixtures

---------
